Create narrow-rail
A narrow double-rail module rail layout with snap to grid enabled.  [Based on Brian's Trains]

Version 1.0 -
Added Madzuri circuit to chest and inserters for even loading to ore provider station
Added 4-Car siding for a smaller siding option
Adjusted U-turn to be a smaller space as it'll strictly be used for U-turns and not intersections
Added modified rail crossing
Added T Intersection to use instead of the U-turn.